More quality metrics

+++ generateSummary.sh
@@ -11,15 +11,19 @@ cp $headerFile $summaryfile
 # This-preprocessing step is necessary to ensure that the results will be reported alphanumerically sorted
 # whereas Nextflow reports them in the order of job completion by default.
 echo Renaming the output files nicely...
-for subdir in `ls */ -d | tr -d '/'`; do
+for subdir in $(ls */ -d | tr -d '/'); do
 	reportFileName=$subdir/report/report.html
-	sampleName=`cat $reportFileName | grep "Sample name:" | awk -F'>' '{ print $4 }' | awk -F'<' '{print $1}'`
-	echo $sampleName
-
-	# Rename the files/folders such that sample name is a prefix
-	for file in `ls $subdir`; do
-		mv $subdir/$file $subdir/${sampleName}_${file}
-	done
+	# Check if this report file exists, else, its contents were already relabelled before
+	if [[ -f $reportFileName ]]; then
+		# Rename the files/folders such that sample name is a prefix
+		sampleName=$(cat $reportFileName | grep "Sample name:" | awk -F'>' '{ print $4 }' | awk -F'<' '{print $1}')
+		for file in $(ls $subdir); do
+			mv $subdir/$file $subdir/${sampleName}_${file}
+		done
+	else
+		sampleName=$(ls $subdir | head -n 1 | awk -F '_' '{print $1}')
+	fi	
+	
 	mv $subdir $sampleName
 done
 
@@ -31,32 +35,41 @@ echo Gathering and tabulating summary statistics...
 echo "<h2>Summary</h2>" >> $summaryfile
 echo '<table>' >> $summaryfile
 echo "<tr>" >> $summaryfile
-echo "<th>Sample#</th><th>Sample name</th><th>Total #reads</th><th>% covid hits</th><th>Genomic coordinates 0X</th><th>Genomic coordinates <10X</th>" >> $summaryfile
+echo "<th>Sample#</th><th>Sample name</th><th>Total #reads</th><th>Reads aligned PF*</th><th>Genomic coordinates 0X</th><th>Genomic coordinates <10X</th>" >> $summaryfile
 echo "</tr>" >> $summaryfile
 
 
+# Also generate a separate csv file containing the summary table 
+csvFile=./summaryTable.csv
+echo "Sample#,Sample name,Total #reads,Reads aligned PF,Genomic coordinates 0X,Genomic coordinates <10X" >> $csvFile
+
 sampleNames=()
 plottingData=()
-for sampleName in `ls */ -d | tr -d '/'`; do
+for sampleName in $(ls */ -d | tr -d '/'); do
 	sampleNames+=($sampleName)
 	
 	# Extract some of the statistics from individual report files
 	reportFileName=$sampleName/${sampleName}_report/report.html
-	numReads=`cat $reportFileName | grep "Total number of reads:" | awk -F'>' '{ print $4 }' | awk -F'<' '{print $1}'`
-	pctCovid=`cat $reportFileName | grep "Hits to SARS-Cov2 genome" | awk -F'(' '{ print $3 }' | awk -F'%' '{ print $1 }'`
-	numUncovered=`cat $reportFileName | grep "ncovered genomic coordinates (0X):" | awk -F'>' '{ print $4 }' | awk '{ print $1 }'`
-	numPoorlyCovered=`cat $reportFileName | grep "poorly covered genomic coordinates (<10X):" | awk -F'>' '{ print $4 }' | awk '{ print $1 }'`
-	plottingData+=($numReads $pctCovid)
+	numReads=$(cat $reportFileName | grep "Total number of reads:" | awk -F '>' '{ print $4 }' | awk -F '<' '{print $1}')
+	readsMapped=$(cat $reportFileName | grep "Reads passing filter" | awk -F '>' '{ print $4 }' | awk -F '<' '{ print $1 }')
+	pctMapped=$(echo $readsMapped | awk '{print $2}' | tr -d '()%')
+	numUncovered=$(cat $reportFileName | grep "All genomic coordinates:" | awk -F '>' '{ print $4 }' | awk -F '<' '{ print $1 }')
+	numPoorlyCovered=$(cat $reportFileName | grep "All genomic coordinates:" | awk -F '>' '{ print $6 }' | awk -F '<' '{ print $1 }')
+	plottingData+=($numReads $pctMapped)
 
-	# Record as a row in the data table
+	# Record as a row in the data table	
 	echo "<tr>" >> $summaryfile
 	echo "<td>${#sampleNames[@]}</td><td><a href=\"./$sampleName/${sampleName}_report/report.html\">$sampleName</a></td>" >> $summaryfile 
-	echo "<td>$numReads</td><td>$pctCovid</td><td>$numUncovered</td><td>$numPoorlyCovered</td>" >> $summaryfile
+	echo "<td>$numReads</td><td>$readsMapped</td><td>$numUncovered</td><td>$numPoorlyCovered</td>" >> $summaryfile
 	echo "</tr>" >> $summaryfile
+	
+	echo "${#sampleNames[@]},$sampleName,$numReads,$readsMapped,$numUncovered,$numPoorlyCovered" >> $csvFile
 done
 
 echo '</table>' >> $summaryfile
 echo "<br><br><br>" >> $summaryfile
+
+echo "*Quantity of raw reads that align to the reference sequence and pass filter, i.e. the read length after adaptor trimming ≥ 30 and minimum read quality ≥ 20 within a sliding window of width 4."
 
 
 # Insert a bar plot comparing covid reads in all samples of this run.
@@ -69,11 +82,11 @@ echo "<br><br><br>" >> $summaryfile
 echo Adding figures
 for ((i = 0; i < ${#sampleNames[@]}; i+=2)); do
 	echo "<span style=\"float:left;\"><a href=\"./${sampleNames[$i]}/${sampleNames[$i]}_report/report.html\">${sampleNames[$i]}</a></span>" >> $summaryfile
-	echo "<span style=\"float:right;\"><a href=\"./${sampleNames[`expr $i + 1`]}/${sampleNames[`expr $i + 1`]}_report/report.html\">" >> $summaryfile
-	echo "${sampleNames[`expr $i + 1`]}</a></span><br>" >> $summaryfile
+	echo "<span style=\"float:right;\"><a href=\"./${sampleNames[$(expr $i + 1)]}/${sampleNames[$(expr $i + 1)]}_report/report.html\">" >> $summaryfile
+	echo "${sampleNames[$(expr $i + 1)]}</a></span><br>" >> $summaryfile
 	
 	echo "<img src=\"./${sampleNames[$i]}/${sampleNames[$i]}_report/quality.png\" width=\"49%\" class=\"center\">" >> $summaryfile
-	echo "<img src=\"./${sampleNames[`expr $i + 1`]}/${sampleNames[`expr $i + 1`]}_report/quality.png\" width=\"49%\" class=\"center\">" >> $summaryfile
+	echo "<img src=\"./${sampleNames[$(expr $i + 1)]}/${sampleNames[$(expr $i + 1)]}_report/quality.png\" width=\"49%\" class=\"center\">" >> $summaryfile
 	echo "<br><br><br>" >> $summaryfile
 done
 	
@@ -83,26 +96,26 @@ echo Appending a list of data analysis parameters...
 echo >> $summaryfile
 echo "<h2>Software configuration</h2>" >> $summaryfile
 
-pangolinVersion=`pangolin -v`
-pangolearnVersion=`pangolin -pv`
+pangolinVersion=$(pangolin -v)
+pangolearnVersion=$(pangolin -pv)
 
 if [[ $platform == "Illumina" ]]; then
-	echo "Bowtie2 v`bowtie2 --version | head -n 1 | awk '{ print $3 }'`, " >> $summaryfile
+	echo "Bowtie2 v)bowtie2 --version | head -n 1 | awk '{ print $3 }'), " >> $summaryfile
 else
-	echo "Minimap2:" `minimap2 --version` >> $summaryfile
+	echo "Minimap2:" $(minimap2 --version) >> $summaryfile
 fi
 
-echo `samtools --version | head -n 2`", " >> $summaryfile
-echo `ivar version | head -n 1`", " >> $summaryfile
-echo `kraken2 -v | head -n 1`", " >> $summaryfile
-echo `kallisto version`"." >> $summaryfile
+echo $(samtools --version | head -n 2)", " >> $summaryfile
+echo $(ivar version | head -n 1)", " >> $summaryfile
+echo $(kraken2 -v | head -n 1)", " >> $summaryfile
+echo $(kallisto version)"." >> $summaryfile
 
-allIncludedLineages=`$rootDir/listVariantsAvail.py $variantDBfile`
-echo "Lineage definitions were compiled on" `date +%Y-%m-%d -r $variantDBfile` \
+allIncludedLineages=$($rootDir/listVariantsAvail.py $variantDBfile)
+echo Lineage definitions were compiled on $(date +%Y-%m-%d -r $variantDBfile) \
 		"from <a href=\"https://github.com/cov-lineages/constellations/tree/main/constellations/definitions\">constellations</a>." >> $summaryfile
-echo "Lineage signature file was compiled on" `date +%Y-%m-%d -r $variantDBfile` \
-		"and includes lineages: $allIncludedLineages." >> $summaryfile
-echo "Lineage assignment to the consensus sequence was performed by $pangolinVersion using the classification tree of $pangolearnVersion." >> $summaryfile
+echo Lineage signature file was compiled on $(date +%Y-%m-%d -r $variantDBfile) \
+		and includes lineages: $allIncludedLineages. >> $summaryfile
+echo Lineage assignment to the consensus sequence was performed by $pangolinVersion using the classification tree of $pangolearnVersion. >> $summaryfile
 	
 echo "</body>" >> $summaryfile
 echo "</html>" >> $summaryfile
